make a hook for jbrowse
